# openapi/openapi-sdtooling-tier1-v1.yaml
info:
  title: SD Tooling API
  description: |-
    The APIs described in the following document enable the management
    of the lifecycle for creating a Resource Description on the provider side, including
    its publication on the federated catalogue. The available APIs allow to define
    access and usage policies, specify the Resource Address on the provider side,
    and manage the validation and publication process on the federated catalogue.

    ## Business Context

    The API facilitates the creation and management of Resource Descriptions in a federated data space. It enables providers to publish metadata about their data assets on a federated catalogue, define access and usage policies, and manage the validation and publication workflow. This supports discoverability and interoperability across the data ecosystem.


    ### Key Actors

    - **Provider**: Creates and publishes Resource Descriptions, defines policies, and manages validation.

    - **Governance Authority**: Hosts the Federated Catalogue

    ### Typical Flow

    1. Create a Resource Description.

    2. Define access and usage policies.

    3. Specify the Resource Address.

    4. Submit the Resource Description for validation

    5. Publish the Resource Description on the federated catalogue
  contact:
    name: Simpl Programme
    url: https://simpl-programme.ec.europa.eu/
    email: cnect-simpl@ec.europa.eu
  license:
    name: European Union Public License (EUPL) 1.2
...
  version: "1.0"
servers:
...
security:
- bearerAuth: []
openapi: 3.0.1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: IAA cloud gateway JWT token
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InputObject:
      type: object
      properties:
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
...
    ResourceAddress:
      required:
      - value
      type: object
      properties:
        templateId:
          type: string
        value:
          type: string
    InputValidationIssue:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/InputObject'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    Template:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
...
    InternalServerErrorProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    NotFoundProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    UnauthorizedProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    BadRequestProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        issues:
          type: array
          items:
            $ref: '#/components/schemas/InputValidationIssue'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
paths:
...
  /resourceAddresses/templates/{templateId}/schema:
    get:
      tags:
      - Resource Address
      summary: Returns a source address template schema based on templateId
      description: Returns a template (JSON schema) describing the fields and constraints to create the sourceAddress.
      operationId: getSourceAddressSchema
      parameters:
      - name: templateId
        in: path
        description: The template id
        required: true
        schema:
          type: string
          example: TPL-1
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Parameter validation error
                    status: 400
                    detail: One or more parameters are invalid
                    instance: /resourceAddresses/templates/{templateId}/schema
                    issues:
                    - type: urn:problem-type:simpl:invalidParameter
                      title: Invalid parameter
                      detail: The templateId parameter must meet specific criteria
                      in: path
                      name: templateId
                      value: invalidValue
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                InternalServerError:
                  summary: Unexpected internal error
                  description: InternalServerError
                  value:
                    type: internal-error
                    title: Unexpected internal error
                    status: 500
                    detail: 'GENERAL_ERROR: Please try again later'
                    instance: /resourceAddresses/templates/{templateId}/schema
        "200":
          description: Template (JSON schema) for the sourceAddress
          content:
            application/json:
              schema:
                type: object
              examples:
                IonosS3:
                  summary: IonosS3 template example
                  description: IonosS3
                  value:
                    type: object
                    properties:
                      bucketName:
                        type: string
                      region:
                        type: string
                    required:
                    - bucketName
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/NotFoundProblem'
              examples:
                NotFound:
                  summary: Resource not found
                  description: NotFound
                  value:
                    type: resource-not-found
                    title: RESOURCE_ADDRESS_NOT_FOUND
                    status: 404
                    detail: UI schema not found
                    instance: /resourceAddresses/templates/{templateId}/schema
  /resourceAddresses/assets/{assetId}:
    get:
      tags:
      - Resource Address
      summary: Retrieves the resource address associated to the given assetId
      description: Retrieves the resource address associated to the given assetId
      operationId: getResourceAddress
      parameters:
      - name: assetId
        in: path
        description: Identifier of asset within connector
        required: true
        schema:
          type: string
          example: 123
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
                    instance: /resourceAddresses/assets/{assetId}
        "200":
          description: Successfully retrieved the resource address associated to the given assetId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceAddress'
              examples:
                Resource Address:
                  description: Resource description address to the given assetId
                  value:
                    templateId: "5"
                    value: '{"type":"MinioS3","endpoint":"https://minio01.integrated.simpl-europe.eu","bucketName":"provider-bucket","objectName":"example-s3.txt"}'
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/UnauthorizedProblem'
              examples:
                Unauthorized:
                  summary: Missing or invalid authorization
                  description: Unauthorized
                  value:
                    type: unauthorized
                    title: Unauthorized
                    status: 401
                    detail: Missing or invalid Authorization header
                    instance: /resourceAddresses/assets/{assetId}
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/NotFoundProblem'
              examples:
                Not found Error Example:
                  summary: AssetId Not Found
                  description: Not found Error Example
                  value:
                    type: urn:problem-type:simpl:resourceAddressNotFound
                    title: Resource Address Not Found
                    status: 404
                    detail: No resource address found for assetId
                    instance: /resourceAddresses/assets/{assetId}
...
  /resourceAddresses/sharingMethods:
    get:
      tags:
      - Resource Address
      summary: Returns the list of sharing methods by offer type.
      description: Returns the list of sharing methods mapped to the required offer type (DATA, APPLICATION, INFRASTRUCTURE).
      operationId: getResourceSharingMethods
      parameters:
      - name: offeringType
        in: query
        description: The offering type identifier
        required: true
        schema:
          type: string
          enum:
          - APPLICATION
          - DATA
          - INFRASTRUCTURE
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Parameter validation error
                    status: 400
                    detail: One or more parameters are invalid
                    instance: /resourceAddresses/sharingMethods
                    issues:
                    - type: urn:problem-type:simpl:invalidParameter
                      title: Invalid parameter
                      detail: The offeringType parameter must         meet specific criteria
                      in: query
                      name: offeringType
                      value: invalidValue
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                InternalServerError:
                  summary: Unexpected internal error
                  description: InternalServerError
                  value:
                    type: internal-error
                    title: Unexpected internal error
                    status: 500
                    detail: 'GENERAL_ERROR: Please try again later'
                    instance: /resourceAddresses/sharingMethods
        "200":
          description: A list of sharing methods
          content:
            application/json:
              schema:
                type: object
              examples:
                validTypes:
                  summary: Valid sharing methods
                  description: validTypes
                  value:
                  - IONOS_S3
                  - HTTPDATA_PUSH
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/UnauthorizedProblem'
              examples:
                Unauthorized:
                  summary: Missing or invalid authorization
                  description: Unauthorized
                  value:
                    type: unauthorized
                    title: Unauthorized
                    status: 401
                    detail: Missing or invalid Authorization header
                    instance: /resourceAddresses/sharingMethods
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/NotFoundProblem'
              examples:
                NotFound:
                  summary: Resource not found
                  description: NotFound
                  value:
                    type: resource-not-found
                    title: RESOURCE_ADDRESS_NOT_FOUND
                    status: 404
                    detail: UI schema not found
                    instance: /resourceAddresses/sharingMethods
  /resourceAddresses/sharingMethods/{sharingMethodId}/templates:
    get:
      tags:
      - Resource Address
      summary: Returns the list of source address templates by sharing method and offering type.
      description: Returns the list of templates mapped to the required sharing method and offering type (DATA, APPLICATION, INFRASTRUCTURE).
      operationId: getSourceAddressTemplates
      parameters:
      - name: offeringType
        in: query
        description: The offering type identifier
        required: true
        schema:
          type: string
          enum:
          - APPLICATION
          - DATA
          - INFRASTRUCTURE
      - name: sharingMethodId
        in: path
        description: The resource sharing method identifier
        required: true
        schema:
          type: string
          enum:
          - HTTPDATA_PUSH
          - IONOS_S3
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Parameter validation error
                    status: 400
                    detail: One or more parameters are invalid
                    instance: /resourceAddresses/sharingMethods/{sharingMethodId}/templates
                    issues:
                    - type: urn:problem-type:simpl:invalidParameter
                      title: Invalid parameter
                      detail: The sharingMethodId parameter must meet specific criteria
                      in: path
                      name: sharingMethodId
                      value: invalidValue
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                InternalServerError:
                  summary: Unexpected internal error
                  description: InternalServerError
                  value:
                    type: internal-error
                    title: Unexpected internal error
                    status: 500
                    detail: 'GENERAL_ERROR: Please try again later'
                    instance: /resourceAddresses/sharingMethods/{sharingMethodId}/templates
        "200":
          description: A list of templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
              examples:
                templates:
                  summary: List of templates
                  description: templates
                  value:
                  - id: TPL-1
                    label: TPL 1
                  - id: TPL-2
                    label: TPL 2
                  - id: TPL-3
                    label: TPL 3
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/UnauthorizedProblem'
              examples:
                Unauthorized:
                  summary: Missing or invalid authorization
                  description: Unauthorized
                  value:
                    type: unauthorized
                    title: Unauthorized
                    status: 401
                    detail: Missing or invalid Authorization header
                    instance: /resourceAddresses/sharingMethods/{sharingMethodId}/templates
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/NotFoundProblem'
              examples:
                NotFound:
                  summary: Resource not found
                  description: NotFound
                  value:
                    type: resource-not-found
                    title: RESOURCE_ADDRESS_NOT_FOUND
                    status: 404
                    detail: Template not found
                    instance: /resourceAddresses/sharingMethods/{sharingMethodId}/templates
...
  /resourceAddresses/templates/{templateId}/uiSchema:
    get:
      tags:
      - Resource Address
      summary: Returns a source address UI schema based in templateId.
      description: Returns a UI schema that defines the constraints for the form representation of the sourceAddress.
      operationId: getSourceAddressUiSchema
      parameters:
      - name: templateId
        in: path
        description: The template id
        required: true
        schema:
          type: string
          example: TPL-1
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Parameter validation error
                    status: 400
                    detail: One or more parameters are invalid
                    instance: /resourceAddresses/templates/{templateId}/uiSchema
                    issues:
                    - type: urn:problem-type:simpl:invalidParameter
                      title: Invalid parameter
                      detail: The templateId parameter must meet specific criteria
                      in: path
                      name: templateId
                      value: invalidValue
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                InternalServerError:
                  summary: Unexpected internal error
                  description: InternalServerError
                  value:
                    type: internal-error
                    title: Unexpected internal error
                    status: 500
                    detail: 'GENERAL_ERROR: Please try again later'
                    instance: /resourceAddresses/templates/{templateId}/uiSchema
        "200":
          description: UI schema for the sourceAddress
          content:
            application/json:
              schema:
                type: object
              examples:
                amazonS3UI:
                  summary: UI schema example
                  description: amazonS3UI
                  value:
                    ui:order:
                    - bucketName
                    - region
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/NotFoundProblem'
              examples:
                NotFound:
                  summary: Resource not found
                  description: NotFound
                  value:
                    type: resource-not-found
                    title: RESOURCE_ADDRESS_NOT_FOUND
                    status: 404
                    detail: UI schema not found
                    instance: /resourceAddresses/templates/{templateId}/uiSchema
...